+++ docs/changelog.md
@@ -1,4 +1,22 @@
 
+# 0.7.0
+
+- Removed `consensus_masking_threshold` option.
+
+- Removed `keep_all_intermediate_files` option.
+
+- Removed `masked_nextclade` option.
+
+- Removed `masked_ivar` option.
+
+- Removed `testbiin` option
+
+- Added unittests for modules `bin.best_reference`, `bin.detect_variants`.
+
+- `runtest` now outputs to the default temporary directory.
+
+- `runtest` now has `compare_TestResults` to validate the observed results with known results.
+
 # 0.6.0
 
 - Added picard to remove duplicates at the alignment level
@@ -8,3 +26,7 @@
 - Removed sampleReports directory
 
 - Options using masked consensus sequences are availble but will still used unmasked consensensus sequence. This option may be updated or removed in the future.
+
+# 0.5.0
+
+- Major rewrite to remove R dependency.